pyproject.toml
[tool.poetry]
name = "binance-kraken-demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
websockets = "^13.1"
asyncio = "^3.4.3"
django-ninja = "^1.3.0"
httpx = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
